+++ packages/eyes-playwright/src/SpecDriver.js
@@ -3,7 +3,7 @@ const {TypeUtils} = require('@applitools/eyes-sdk-core')
 // #region HELPERS
 
 async function handleToObject(handle) {
-  const type = handle._objectType
+  const [_, type] = handle.toString().split('@')
   if (type === 'array') {
     const map = await handle.getProperties()
     return Promise.all(Array.from(map.values(), handleToObject))
@@ -32,9 +32,6 @@ function transformSelector(selector) {
 
 // #region UTILITY
 
-function isStateless() {
-  return true
-}
 function isDriver(page) {
   return page.constructor.name === 'Page'
 }
@@ -43,21 +40,10 @@ function isElement(element) {
   return element.constructor.name === 'ElementHandle'
 }
 function isSelector(selector) {
-  return TypeUtils.isString(selector)
+  return TypeUtils.has(selector, ['type', 'selector']) || TypeUtils.isString(selector)
 }
 function extractContext(page) {
   return page.constructor.name === 'Page' ? page.mainFrame() : page
-}
-function toEyesSelector(selector) {
-  if (!TypeUtils.isString(selector)) return {selector}
-  if (selector.includes('>>')) return {selector}
-  if (selector.startsWith('//')) return {type: 'xpath', selector}
-
-  const match = selector.match(/(css:light|xpath)=(.+)/)
-  if (!match) return {selector}
-  const [_, type, value] = match
-  if (type === 'css:light') return {type: 'css', selector: value}
-  else if (type === 'xpath') return {type: 'xpath', selector: value}
 }
 function isStaleElementError(err) {
   return err && err.message && err.message.includes('Protocol error (DOM.describeNode)')
@@ -79,7 +65,11 @@ async function executeScript(frame, script, arg) {
 }
 async function mainContext(frame) {
   frame = extractContext(frame)
-  return frame._page.mainFrame()
+  let mainFrame = frame
+  while (mainFrame.parentFrame()) {
+    mainFrame = mainFrame.parentFrame()
+  }
+  return mainFrame
 }
 async function parentContext(frame) {
   frame = extractContext(frame)
@@ -110,6 +100,11 @@ async function getTitle(page) {
 async function getUrl(page) {
   return page.url()
 }
+async function getDriverInfo(_page) {
+  return {
+    // isStateless: true,
+  }
+}
 async function visit(page, url) {
   return page.goto(url)
 }
@@ -124,6 +119,18 @@ async function type(_frame, element, keys) {
 }
 async function waitUntilDisplayed(frame, selector) {
   return frame.waitForSelector(selector)
+}
+async function scrollIntoView(frame, element, align = false) {
+  if (isSelector(element)) {
+    element = await findElement(frame, element)
+  }
+  await frame.evaluate(([element, align]) => element.scrollIntoView(align), [element, align])
+}
+async function hover(frame, element, {x = 0, y = 0} = {}) {
+  if (isSelector(element)) {
+    element = await findElement(frame, element)
+  }
+  await element.hover({position: {x, y}})
 }
 
 // #endregion
@@ -156,10 +163,8 @@ async function build(env) {
     driver = await launcher.launch(options)
   }
   const context = await driver.newContext(device ? playwright.devices[device] : {})
-  return context.newPage()
-}
-async function cleanup(page) {
-  return page && page.context()._browserBase.close()
+  const page = await context.newPage()
+  return [page, () => driver.close()]
 }
 
 // #endregion
@@ -171,7 +176,6 @@ exports.isSelector = isSelector
 exports.extractContext = extractContext
 exports.isStaleElementError = isStaleElementError
 exports.isEqualElements = isEqualElements
-exports.toEyesSelector = toEyesSelector
 
 exports.executeScript = executeScript
 exports.mainContext = mainContext
@@ -184,11 +188,13 @@ exports.getViewportSize = getViewportSize
 exports.setViewportSize = setViewportSize
 exports.getTitle = getTitle
 exports.getUrl = getUrl
+exports.getDriverInfo = getDriverInfo
 exports.visit = visit
 exports.takeScreenshot = takeScreenshot
 exports.click = click
 exports.type = type
 exports.waitUntilDisplayed = waitUntilDisplayed
+exports.scrollIntoView = scrollIntoView
+exports.hover = hover
 
 exports.build = build
-exports.cleanup = cleanup