+++ packages/visual-grid-client/CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Unreleased
 
 
+## 15.0.5 - 2020/9/17
+
+- updated to @applitools/eyes-sdk-core@12.2.5 (from 12.2.4)
+
 ## 15.0.4 - 2020/9/15
 
 - updated to @applitools/eyes-sdk-core@12.2.4 (from 12.2.3)